Prevent several code mistakes (#298)
* Use using so AesCryptoServiceProvider gets properly disposed

* Socket is null so RemoteEndPoint would end in a NullReferenceException

* Fixed possible NullReferenceException during experience table loading.

* Properly dispose FileTokenScanner when a DefineFile gets disposed.

* Fixed possible NullReferenceException during DefineFile reading

* Fixed possible NullReferenceException during WldFile reading

* Prevent NullReferenceException in case size is null

* Properly dispose _reader while disposing a ResourceTableFile

* Don't call Socket.RemoteEndPoint while Socket is null

* Fixed NullReferenceException in Vector3 comparison

* Fix possible NullReferenceException

* Implement GetHashCode for QuestInfo and catch NullReferenceException in Equals function

* Prevent NullReferenceException and removed multiple enumeration.

* Use Int64 arithmetic instead of Int32 arithmetic to prevent an overflow

* Added TODO comment to not executed for loop

* Prevent NullRefenceExceptions in TaskbarSystem

* Simplified if statement and removed multiple enumeration

* A byte can't be < 0

* Don't compare Type twice

* Prevent possible NullReferenceException

* Prevent possible NullReferenceException.

* Prevent possible NullReferenceException.

* Check for null before using response

* Make LoginHandlerTest disposable because it has disposable members

* settingsBlock.GetInstructions can't return null and LINQ's Select and OrderBy don't care about empty IEnumerables.

* GetInstructions can't return null and removed multiple enumeration.

* quest.StartJobs can't be null at that point

* Monsters can't be null at this point

* inventoryItem can't be null at that point

* quests can't be null at that point.

* dialogs can't be null at that point

* Prevent NullReferenceException

* Guarantee that IStatement can be casted to Block.

* Don't try to pick up target when it's not an IItemEntity.

* Prevent NullReferenceException

* Prevent NullReferenceException

* Prevent NullReferenceException

* Prevent NullReferenceException

* Prevent NullReferenceException

* Removed unneeded null checks and check null before using line

* Use pattern matching

* Change SfxIdPacket to new packet structure

* Changed SfxIdPacket type to class

* Use types instead of var

* Call _player.Follow.Reset() instead of returning

* Use pattern matching

* Use type instead of var

* Removed left over apostrophe.

+++ src/Rhisis.World/Systems/PlayerData/PlayerDataSystem.cs
@@ -26,7 +26,8 @@ namespace Rhisis.World.Systems.PlayerData
         /// <inheritdoc />
         public bool IncreaseGold(IPlayerEntity player, int goldAmount)
         {
-            long gold = player.PlayerData.Gold + goldAmount;
+            // We cast player gold to long because otherwise it would use Int32 arithmetic and would overflow
+            long gold = (long)player.PlayerData.Gold + goldAmount;
 
             if (gold > int.MaxValue || gold < 0) // Check gold overflow
             {
